+++ Nagisa/SettingsDialog.xaml.cpp
@@ -1,0 +1,25 @@
+﻿//
+// SettingsDialog.xaml.cpp
+// SettingsDialog 类的实现
+//
+
+#include "pch.h"
+#include "SettingsDialog.xaml.h"
+
+using namespace Nagisa;
+
+using namespace Platform;
+using namespace Windows::Foundation;
+using namespace Windows::Foundation::Collections;
+using namespace Windows::UI::Xaml;
+using namespace Windows::UI::Xaml::Controls;
+using namespace Windows::UI::Xaml::Controls::Primitives;
+using namespace Windows::UI::Xaml::Data;
+using namespace Windows::UI::Xaml::Input;
+using namespace Windows::UI::Xaml::Media;
+using namespace Windows::UI::Xaml::Navigation;
+
+SettingsDialog::SettingsDialog()
+{
+	InitializeComponent();
+}